Updated gopro meta extraction to take cmd line args. This allows it to be called from the python image extraction tool.
